requirements.txt
alembic==0.8.6
aniso8601==1.0.0
coverage==4.0.1
fake-factory==0.5.3
Flask==0.10.1
Flask-API==0.6.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PyJWT==1.4.0
pytz==2015.6
six==1.10.0
SQLAlchemy==1.0.8
Werkzeug==0.10.4
wheel==0.24.0
flask-restful==0.3.5